WIP register old stuff
