+++ test-apps/appui-test-app/appui-test-providers/src/appui-test-providers.ts
@@ -3,7 +3,16 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from "./tools/ContentLayoutTools";
+export * from "./tools/GenericLocateTool";
+export * from "./tools/InspectUiItemInfoTool";
+export * from "./tools/OpenAbstractModalDialogTool";
+export * from "./tools/OpenCustomDialogTool";
+export * from "./tools/OpenPopoutDialogTool";
+export * from "./tools/OpenPopoutViewTool";
+export * from "./tools/OpenSynchronizedViewTool";
 export * from "./tools/RegisterUiProviderTool";
+export * from "./tools/SampleTool";
+export * from "./tools/ToolWithDynamicSettings";
 export * from "./tools/UiLayoutTools";
 
 export * from "./ui/components/LanguageSelect";
@@ -11,6 +20,8 @@ export * from "./ui/components/LanguageSelect";
 export * from "./ui/dialogs/SampleModalDialog";
 
 export * from "./ui/providers/AbstractUiItemsProvider";
+export * from "./ui/providers/ContentLayoutStageUiItemsProvider";
+export * from "./ui/providers/CustomContentStageUiProvider";
 export * from "./ui/providers/FloatingWidgetsUiItemsProvider";
 export * from "./ui/providers/InspectUiItemInfoToolProvider";
 export * from "./ui/providers/MessageUiItemsProvider";
@@ -19,8 +30,8 @@ export * from "./ui/providers/UpdatedUiItemsProvider";
 export * from "./ui/providers/WidgetContentProvider";
 
 export * from "./ui/frontstages/ComponentExamples";
-export * from "./ui/frontstages/ContentLayout";
-export * from "./ui/frontstages/CustomContent";
+export * from "./ui/frontstages/ContentLayoutFrontstage";
+export * from "./ui/frontstages/CustomContentFrontstage";
 export * from "./ui/frontstages/CustomFrontstageProvider";
 export * from "./ui/frontstages/PopoutWindowsFrontstage";
 export * from "./ui/frontstages/registerCustomFrontstage";
